requirements.txt
# Core dependencies
mcp>=1.0.0
fastmcp>=0.1.0
mem0ai>=0.1.0
python-dotenv>=1.0.0

# Server dependencies
uvicorn[standard]>=0.30.0
starlette>=0.37.0

# Utilities
aiohttp>=3.9.0

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# ruff>=0.1.0